Get frequency distribution of NA values per row
Currently, we obtain multiple appointments for each patient by
repeatedly querying the appointments table. For convenience, we have set
the number of queries to 10. To determine whether 10 is a reasonable
number of queries, we get the frequency distribution of the count of NA
values per row. If one or more rows have zero NA values, then we should
increase the number of queries.

Closes #2