Refactor `reshape_dataset`
The most important part of `reshape_dataset` is the code that stacks the
input, wide dataset into the output, long dataset. We move this code
into the `stacker` function, and add a test. We also move the
configuration code (`f_in`, `f_out`, `index_names`, etc.) into `main`.
If we add variables to the study, and they shouldn't be stacked, then we
add them to `index_names`; if they should be stacked, then we increment
`group_size`.