Add `appointments_generate_dataset_sql` action
This is an SQL Runner action that extracts a similar dataset to the
dataset definition, albeit in "long" rather than "wide" form. By using
SQL Runner we (hopefully) overcome some of the limitations of Data
Builder.

To make comparing the query to the dataset definition easier, we rename
a couple of variables in the dataset definition. Otherwise, the dataset
definition is unchanged.