Calculating number of unique patients over time
This calculates the number of unique patients for each practice in each month. We generate the decile chart for interest, though, like those for the number of appointments, this should be interpreted with caution as practices will vary in size. It should tell us whether the number of unique patients is increasing over time, generally speaking.