Calculating proportions of lead times
This includes a new measure calculation, which measures the proportion of appointments that occur within a specific number of days (this specific number is passed to the script as a variable). Specifically, we are interested in calculating the proportion of appointments that (i) occur on the day that they were booked (i.e., lead time = 0 days) and (ii) occur within two days of the booking (i.e., lead time <=2 days. Actions hve been added to the project.yaml to capture these calculations (using both the booked and start date as the x axis anchor).