Removing practice ID from measure output
Looking at the output on the server, the values for the proportion within 0/2 days are not bounded by 0 and 1. I'm wondering whether this is because practice ID is being read as either the numerator or denominator. @iaindilligham did mention this previously - I did run a version of the pipeline with and without practice ID locally, which didn't seem to make a difference to the decile plots that were generated. I have also tried to reun a version that has more realistic practice IDs (i.e., not 1/2/3 but 1000, 2000, 3000) but this didn't seem to make any difference.

In any case, this commit removes practice from the measures file, to see if the output is more sensible.