Amending measures output for deciles-chart action
The deciles chart action requires to columns to be present - value and date. I previously had thought that three columns were required - population, value and date. This commit reworks all measure files such that the population column is not included. Most of these measures files were using a dummy value of 1 for the population anyway, but the correct calculation needed to be added to the `generate_proportion_lead_time_measure.py` script.