Adding a function (+ tests) to calculate summaries by season
The summarise_to_seasons() function takes a dataframe and, given the necessary specifications of columns and summary function, returns data that has been summarised to seasons. We are intending to use the summary functions np.sum and np.median, but others could be passed. Also included in this commit is a test for this function.