Calculating seasonal summaries for the appointment measures
All appointment measures have been updated to generate two output files. The first is a monthly file that is used to generate the decile plots. The second is a seasonal summary file that will be used to generate seasonal differences for all metrics. The summarise_to_seaons() function is used throughout to generate the summary statistics. The project.yaml file has been updated accordingly.