Initialising a preliminary report
An RMarkdown document has been added to the pipeline so that outputs and results can be compiled over time. This currently outputs as a Powerpoint file so as to allow offline changes. Whether the action will be run on the server or run offline once outputs are released is not yet decided.