Filtering appointments data to a date range
The appointments data is generated differently to the rest of the data. When calculating the seasonal summaries, we need to filter the appointments data to use the same time frame as used elsewhere.

We provide the date range (`--start-date` and `--end-date`) as an argument to the `generate_appointment_seasonal_summaries.py` script in `project.yaml`. These are not required arguments, and if not provided, the filtering defaults to `1900-01-01` and `date.today()`.

Note that monthly appointments data need not be filtered because we need the longer study period to generate fully populated decile plots.