Recording where we might lose practices due to merging
Previously we used a `left_join` but we don't know if that will create NAs in our dataset where practices are not represented in both the (1) raw appointment/patient counts *and* (2) the monthly practice populations. This records data for practices that are not common to both in `output/check/*CHECK.csv` files (as these files need to be `moderately_sensitive` the practice ID has been removed). Tests have been expanded as well.