Fixing a bug
When generating the report locally I noticed that no all the measures were included in the summary statistics for the seasonal comparisons (generated by the `combined_seasonal_data` action). I chased this around a bit and in doing so have:

1. set `row.names=FALSE` in `combine_seasonal_data.R`
2. amended the corresponding `read_csv()` command in `summarise_seasonal_comparisons.R`
3. added some reporting lines in: `combined_seasonal_data.R`

I eventually realised that the problem was due to two things:
1. applying the function `create_seasonal_kids_plots()` to the appointment data (and a resulting 'File not found' type error)
2. the `tryCatch()` that was added previously to manage the dummy data.

It seems like the tryCatch was not generating all necessary output files. I'm not sure why this is, but as this is urgent, I have reinstated @cewalters previous methodology using `lapply`. This will cause a :x: in tests, but we knew that was going to happen anyway.