Redacting histogram counts
Retaining the middle 95% of the data was not enough to ensure that we didn't have small number counts in our histograms. This uses the same `redact_and_round()` function redact the histogram counts and generate a histogram from those redacted counts. We have to do a lot of checking of `nrows()` to deal with dummy data.