Update to seasonal calculations (with test change)
Here we modify the `generate_appointment_seasonal_summaries.py` script to use the study dates as store in the `.json` file. This will truncate the data to the minimum and maximum date as represented across ALL of the definitions in the `.json` file. The previous seasonal mapping is commented out while we develop this new methodology.

The tests have been updated so as to be appropriate for the new seasonal definitions. We should be aware of this for when definitions of the seasons change in the future.